Edited JavaFX
Added more dyanmic colours for different cases.
Refactored some parts of code.
