+++ build.gradle
@@ -2,6 +2,7 @@ plugins {
     id 'java'
     id 'application'
     id 'com.github.johnrengelman.shadow' version '7.1.2'
+//    id 'checkstyle'
 }
 
 repositories {
@@ -42,7 +43,7 @@ test {
 }
 
 application {
-    mainClass.set("Main/DukeGPT")
+    mainClass.set("ui/Launcher")
 }
 
 shadowJar {
@@ -53,3 +54,7 @@ shadowJar {
 run{
     standardInput = System.in
 }
+
+//checkstyle {
+//    toolVersion = "11.0.0"
+//}
