+++ data/duke.txt
@@ -1,9 +1,13 @@
-T | 0 | read book
 T | 1 | join sports club
 D | 0 | return thing | 2023-09-01
-E | 1 | testing readToFile | 2025-09-01 | 2025-09-01
 E | 0 | testing readToFile | 2025-09-01 | 2025-09-01
 E | 0 | testing readToFile | 2025-09-01 | 2025-09-01
-D | 0 | this fucking iP | 2025-12-12
+D | 1 | this fucking iP | 2025-12-12
 E | 0 | my funeral | 2025-12-12 | 2025-12-12
 T | 0 | this stupid level 9
+D | 0 | upload | 2025-09-05
+T | 0 | whatever
+T | 0 | 23321
+T | 0 | 3
+T | 0 | wash
+T | 0 | dawiooabnid
