--- src/main/java/Command/Add.java
@@ -1,47 +1,0 @@
-package Command;
-
-import Exception.DukeException;
-import Storage.Storage;
-import Ui.Ui;
-import TaskList.Task;
-
-import java.util.List;
-
-/**
- * Represents an Add command, that adds a {@link Task} to the task list.
- */
-public class Add implements Command {
-    /**
-     * The task to be added
-     */
-    private final Task task;
-
-    /**
-     * Initialises command object with corresponding task
-     * @param task Task to be added
-     */
-    public Add(Task task) {
-        this.task = task;
-    }
-
-    /**
-     * Executes the command, by adding the appropriate task to the list in storage
-     * @param storage   Storage currently being used
-     * @throws DukeException    Handles error during execution
-     */
-    @Override
-    public void execute(Storage storage) throws DukeException {
-        List<Task> t = storage.getTasks();
-        t.add(this.task);
-        storage.readToFile();
-        Ui.display("Added ", this.task.toString(), "Now you have " + t.size() + " tasks in the list.");
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (o instanceof Add c) {
-            return this.task.equals(c.task);
-        }
-        return false;
-    }
-}
