--- src/main/java/Command/Command.java
@@ -1,8 +1,0 @@
-package Command;
-
-import Exception.DukeException;
-import Storage.Storage;
-
-public interface Command {
-    public abstract void execute(Storage storage) throws DukeException;
-}
