--- src/main/java/Exception/DukeException.java
@@ -1,12 +1,0 @@
-package Exception;
-
-import Ui.Ui;
-
-/**
- * Represents custom exception for DukeGpt. Can be extended for more specific functionalities
- */
-public class DukeException extends Exception{
-    public DukeException(String message) {
-        super(message);
-    }
-}
