+++ src/main/java/command/Command.java
@@ -1,0 +1,8 @@
+package command;
+
+import exception.DukeException;
+import storage.Storage;
+
+public interface Command {
+    public abstract String[] execute(Storage storage) throws DukeException;
+}
