+++ src/main/java/command/Exit.java
@@ -1,0 +1,27 @@
+package command;
+
+import exception.DukeException;
+import storage.Storage;
+import ui.Ui;
+
+/**
+ * Represents an Exit command, that closes the program
+ */
+public class Exit implements Command {
+    /**
+     * Execute the command, by displaying message and saving to file
+     * @param storage   storage current being used
+     * @throws DukeException    Handles error during execution
+     */
+    @Override
+    public String[] execute(Storage storage) throws DukeException {
+        Ui.display("That was fun! Hope to see you again!");
+        storage.readToFile();
+        return new String[]{"That was fun! Hope to see you again!"};
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        return o instanceof Exit;
+    }
+}
