+++ src/main/java/command/Greet.java
@@ -1,0 +1,26 @@
+package command;
+
+import exception.DukeException;
+import storage.Storage;
+import ui.Ui;
+
+/**
+ * Represents a Greet command
+ */
+public class Greet implements Command {
+    /**
+     * Execute the command, by displaying message
+     * @param storage
+     * @throws DukeException
+     */
+    @Override
+    public String[] execute(Storage storage) throws DukeException {
+        Ui.display("Hi! I'm main.DukeGpt! Nice to meet you!");
+        return new String[]{"Hi! I'm main.DukeGpt! Nice to meet you!"};
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        return o instanceof Greet;
+    }
+}
