+++ src/main/java/task/ToDo.java
@@ -1,0 +1,25 @@
+package task;
+
+public class ToDo extends Task{
+    public ToDo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+
+    @Override
+    public String toFileString() {
+        return "T" + super.toFileString();
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (o instanceof ToDo t) {
+            return super.equals(o);
+        }
+        return false;
+    }
+}
