+++ src/main/java/ui/MainWindow.java
@@ -1,0 +1,60 @@
+package ui;
+
+import command.Greet;
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.image.Image;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.VBox;
+import main.DukeGpt;
+
+/**
+ * Controller for the main GUI.
+ */
+public class MainWindow extends AnchorPane {
+    @FXML
+    private ScrollPane scrollPane;
+    @FXML
+    private VBox dialogContainer;
+    @FXML
+    private TextField userInput;
+    @FXML
+    private Button sendButton;
+
+    private DukeGpt duke;
+
+    private Image userImage = new Image(this.getClass().getResourceAsStream("/images/user.png"));
+    private Image dukeImage = new Image(this.getClass().getResourceAsStream("/images/dukeGpt.png"));
+
+    @FXML
+    public void initialize() {
+        scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
+    }
+
+    /** Injects the Duke instance */
+    public void setDuke(DukeGpt d) {
+        this.duke = d;
+        String[] response = this.duke.getResponse("greet");
+        dialogContainer.getChildren().addAll(
+                DialogBox.getDukeDialog(response, dukeImage, "Greet")
+        );
+    }
+
+    /**
+     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
+     * the dialog container. Clears the user input after processing.
+     */
+    @FXML
+    private void handleUserInput() {
+        String input = userInput.getText();
+        String[] response = duke.getResponse(input);
+        String commandType = duke.getCommandType();
+        dialogContainer.getChildren().addAll(
+                DialogBox.getUserDialog(new String[]{input}, userImage),
+                DialogBox.getDukeDialog(response, dukeImage, commandType)
+        );
+        userInput.clear();
+    }
+}
