+++ src/main/resources/css/dialog-box.css
@@ -1,0 +1,46 @@
+.label {
+    background:white;
+    -fx-background-color: background;
+    -fx-background-color: linear-gradient(to bottom right, #00ffbf, #00ddff);
+    -fx-border-color: #009e73 #009e73 #11af84 #11af84;
+    -fx-border-width: 2px;
+    -fx-background-radius: 1em 1em 0 1em;
+    -fx-border-radius: 1em 1em 0 1em;
+    -fx-padding: 6px;
+    -fx-border-insets: 0px 7px 0px 7px;
+    -fx-background-insets: 0px 7px 0px 7px;
+    -fx-text-fill: ladder(background, white 49%, black 50%);
+}
+
+
+.reply-label {
+    -fx-background-radius: 1em 1em 1em 0;
+    -fx-border-radius: 1em 1em 1em 0;
+}
+
+#displayPicture {
+    /* Shadow effect on image. */
+    -fx-effect: dropshadow(gaussian, rgba(0, 0, 0, 0.2), 10, 0.5, 5, 5);
+
+    /* Change size of image. */
+    -fx-scale-x: 1;
+    -fx-scale-y: 1;
+
+    /* Rotate image clockwise by degrees. */
+    -fx-rotate: 0;
+}
+
+.add-label {
+    -fx-background-color: palegreen;
+    -fx-border-color : derive(palegreen, 80%)
+}
+
+.marked-label {
+    -fx-background-color: forestgreen;
+    -fx-border-color : derive(forestgreen, 80%)
+}
+
+.delete-label {
+    -fx-background-color: lightpink;
+    -fx-border-color : derive(lightpink, 80%)
+}
